+++ CHANGES.md
@@ -1,5 +1,11 @@
 #### [unreleased]
 
+#### 2.4.0 / 2024-08-06
+* convert to only use `iframe` and use Google Doc Viewer for Chrome and mobile
+* dynamically render `iframe` in block based on browser
+* mitigate [Chromium issue](https://issues.chromium.org/issues/40063550) by rendering PDF with Google Doc Viewer
+* thanks to @costdev for helping with dynamic block rendering
+
 #### 2.3.1 / 2023-09-10
 * update GA
 * update tested to